requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn
scikit-image
joblib
tabulate
pytest
